requirements.txt
streamlit
pandas
matplotlib.pyplot
numpy
seaborn
plotly.graph_objects
plotly.offline
plotly.express